added HOC withLoader
