+++ src/HOC/withLoader.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const withLoader = EnhancedComponent => props => (
+  props.isLoading
+    ? <div>Загрузка...</div>
+    : <EnhancedComponent {...props} />
+);
+
+export default withLoader;
